# swagger.yaml
openapi: 3.0.3
info:
  title: Challenge Chapter 5
  description: |-
    Welcome to My Home
  contact:
    email: galuhraka18@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /auth/authenticate:
    get:
      tags:
        - Authentication
      operationId: authenticate
      summary: Authenticate a user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successfull operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/UserModel"
                          - type: object
                          - properties:
                              profile:
                                $ref: "#/components/schemas/ProfileModel"

        "400":
          description: Failed operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BadRequestResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/UnauthorizedResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UnauthorizedResponse:
      title: UnauthorizedResponse
      type: object
      properties:
        status:
          type: boolean
          example: false
        message:
          type: string
          example: "Unauthorized"
        data:
          example: null
    BasicResponse:
      title: BasicResponse
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: "success"
    BadRequestResponse:
      title: BadRequestResponse
      type: object
      properties:
        status:
          type: boolean
          example: false
        message:
          type: string
          example: "some error from user"
        data:
          example: null
...
    UserModel:
      title: User
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "galuhrakamahendra"
        email:
          type: string
          example: "galuhrakamahendra@mail.com"
...
    ProfileModel:
      title: Profile
      type: object
      properties:
        identity_type:
          type: string
          example: "NO HP"
        identity_number:
          type: string
          example: "08123456789"
        address:
          type: string
          example: "Solo"
...